New training files created
